refactor: Improve if statement for the filter
Previously for the filter of month and day, "if" part is followed
by "elif" with a list that will assign month and day variable to "all".
However, since the list following "if" and the list following "elif"
are mutually exclusive. In another word, there is not need to
explicitly state the list. Instead, using "else" is enough.